+++ test/test_GMRES/test_GMRES_solve_sgl.cpp
@@ -19,8 +19,8 @@ class GMRESSingleSolveTest: public TestBase {};
 TEST_F(GMRESSingleSolveTest, SolveConvDiff64) {
 
     constexpr int n(64);
-    Matrix<float, n, n> A(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_64_A.csv"));
-    Matrix<float, n, 1> b(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_64_b.csv"));
+    Matrix<double, n, n> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_A.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_b.csv"));
     GMRESSolve<float> gmres_solve_s(A, b, u_sgl, n, conv_tol_sgl);
 
     gmres_solve_s.solve();
@@ -34,8 +34,8 @@ TEST_F(GMRESSingleSolveTest, SolveConvDiff64) {
 TEST_F(GMRESSingleSolveTest, SolveConvDiff256) {
 
     constexpr int n(256);
-    Matrix<float, Dynamic, Dynamic> A(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_256_A.csv"));
-    Matrix<float, n, 1> b(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_256_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_256_A.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_256_b.csv"));
     GMRESSolve<float> gmres_solve_s(A, b, u_sgl, n, conv_tol_sgl);
 
     gmres_solve_s.solve();
@@ -49,8 +49,8 @@ TEST_F(GMRESSingleSolveTest, SolveConvDiff256) {
 TEST_F(GMRESSingleSolveTest, SolveConvDiff1024_LONGRUNTIME) {
 
     constexpr int n(1024);
-    Matrix<float, Dynamic, Dynamic> A(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_1024_A.csv"));
-    Matrix<float, n, 1> b(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_1024_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_1024_A.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_1024_b.csv"));
     GMRESSolve<float> gmres_solve_s(A, b, u_sgl, n, conv_tol_sgl);
 
     gmres_solve_s.solve();
@@ -64,8 +64,8 @@ TEST_F(GMRESSingleSolveTest, SolveConvDiff1024_LONGRUNTIME) {
 TEST_F(GMRESSingleSolveTest, SolveRand20) {
 
     constexpr int n(20);
-    Matrix<float, n, n> A(read_matrix_csv<float>(solve_matrix_dir + "A_20_rand.csv"));
-    Matrix<float, n, 1> b(read_matrix_csv<float>(solve_matrix_dir + "b_20_rand.csv"));
+    Matrix<double, n, n> A(read_matrix_csv<double>(solve_matrix_dir + "A_20_rand.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "b_20_rand.csv"));
     GMRESSolve<float> gmres_solve_s(A, b, u_sgl, n, conv_tol_sgl);
 
     gmres_solve_s.solve();
@@ -79,8 +79,8 @@ TEST_F(GMRESSingleSolveTest, SolveRand20) {
 TEST_F(GMRESSingleSolveTest, Solve3Eigs) {
 
     constexpr int n(25);
-    Matrix<float, n, n> A(read_matrix_csv<float>(solve_matrix_dir + "A_25_3eigs.csv"));
-    Matrix<float, n, 1> b(read_matrix_csv<float>(solve_matrix_dir + "b_25_3eigs.csv"));
+    Matrix<double, n, n> A(read_matrix_csv<double>(solve_matrix_dir + "A_25_3eigs.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "b_25_3eigs.csv"));
     GMRESSolve<float> gmres_solve_s(A, b, u_sgl, n, conv_tol_sgl);
 
     gmres_solve_s.solve();
@@ -95,8 +95,8 @@ TEST_F(GMRESSingleSolveTest, Solve3Eigs) {
 TEST_F(GMRESSingleSolveTest, DivergeBeyondSingleCapabilities) {
 
     constexpr int n(64);
-    Matrix<float, n, n> A(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_64_A.csv"));
-    Matrix<float, n, 1> b(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_64_b.csv"));
+    Matrix<double, n, n> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_A.csv"));
+    Matrix<double, n, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_b.csv"));
 
     // Check convergence under single capabilities
     GMRESSolve<float> gmres_solve_s(A, b, u_sgl, n, conv_tol_sgl);
