+++ test/test_SOR/test_SOR.cpp
@@ -71,8 +71,8 @@ TEST_F(SORTest, SolveConvDiff256_Double_LONGRUNTIME) {
 
 TEST_F(SORTest, SolveConvDiff64_Single) {
     
-    Matrix<float, Dynamic, Dynamic> A(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_64_A.csv"));
-    Matrix<float, Dynamic, 1> b(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_64_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_A.csv"));
+    Matrix<double, Dynamic, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_b.csv"));
 
     for (auto w = ws.cbegin(); w != ws.cend(); ++w) {
 
@@ -91,8 +91,8 @@ TEST_F(SORTest, SolveConvDiff64_Single) {
 
 TEST_F(SORTest, SolveConvDiff256_Single_LONGRUNTIME) {
     
-    Matrix<float, Dynamic, Dynamic> A(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_256_A.csv"));
-    Matrix<float, Dynamic, 1> b(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_256_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_256_A.csv"));
+    Matrix<double, Dynamic, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_256_b.csv"));
 
     for (auto w = ws.cbegin(); w != ws.cend(); ++w) {
 
@@ -111,8 +111,8 @@ TEST_F(SORTest, SolveConvDiff256_Single_LONGRUNTIME) {
 
 TEST_F(SORTest, SolveConvDiff64_SingleFailBeyondEpsilon) {
     
-    Matrix<float, Dynamic, Dynamic> A(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_64_A.csv"));
-    Matrix<float, Dynamic, 1> b(read_matrix_csv<float>(solve_matrix_dir + "conv_diff_64_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_A.csv"));
+    Matrix<double, Dynamic, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_b.csv"));
 
     for (auto w = ws.cbegin(); w != ws.cend(); ++w) {
 
@@ -131,14 +131,14 @@ TEST_F(SORTest, SolveConvDiff64_SingleFailBeyondEpsilon) {
 
 TEST_F(SORTest, SolveConvDiff64_Half) {
     
-    Matrix<half, Dynamic, Dynamic> A(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_64_A.csv"));
-    Matrix<half, Dynamic, 1> b(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_64_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_A.csv"));
+    Matrix<double, Dynamic, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_b.csv"));
 
     for (auto w = ws.cbegin(); w != ws.cend(); ++w) {
 
         cout << "Testing w=" << *w << endl;
 
-        SORSolve<half> SOR_solve_h(A, b, static_cast<half>(*w), max_iter, conv_tol_hlf);
+        SORSolve<half> SOR_solve_h(A, b, *w, max_iter, conv_tol_hlf);
         SOR_solve_h.solve();
         if (show_plots) { SOR_solve_h.view_relres_plot("log"); }
         
@@ -151,14 +151,14 @@ TEST_F(SORTest, SolveConvDiff64_Half) {
 
 TEST_F(SORTest, SolveConvDiff256_Half_LONGRUNTIME) {
     
-    Matrix<half, Dynamic, Dynamic> A(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_256_A.csv"));
-    Matrix<half, Dynamic, 1> b(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_256_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_256_A.csv"));
+    Matrix<double, Dynamic, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_256_b.csv"));
 
     for (auto w = ws.cbegin(); w != ws.cend(); ++w) {
 
         cout << "Testing w=" << *w << endl;
 
-        SORSolve<half> SOR_solve_h(A, b, static_cast<half>(*w), max_iter, conv_tol_hlf);
+        SORSolve<half> SOR_solve_h(A, b, *w, max_iter, conv_tol_hlf);
         SOR_solve_h.solve();
         if (show_plots) { SOR_solve_h.view_relres_plot("log"); }
         
@@ -171,14 +171,14 @@ TEST_F(SORTest, SolveConvDiff256_Half_LONGRUNTIME) {
 
 TEST_F(SORTest, SolveConvDiff64_HalfFailBeyondEpsilon) {
     
-    Matrix<half, Dynamic, Dynamic> A(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_64_A.csv"));
-    Matrix<half, Dynamic, 1> b(read_matrix_csv<half>(solve_matrix_dir + "conv_diff_64_b.csv"));
+    Matrix<double, Dynamic, Dynamic> A(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_A.csv"));
+    Matrix<double, Dynamic, 1> b(read_matrix_csv<double>(solve_matrix_dir + "conv_diff_64_b.csv"));
 
     for (auto w = ws.cbegin(); w != ws.cend(); ++w) {
 
         cout << "Testing w=" << *w << endl;
 
-        SORSolve<half> SOR_solve_h(A, b, static_cast<half>(*w), fail_iter, 0.1*u_hlf);
+        SORSolve<half> SOR_solve_h(A, b, *w, fail_iter, 0.1*u_hlf);
         SOR_solve_h.solve();
         if (show_plots) { SOR_solve_h.view_relres_plot("log"); }
